Add useSelectors hook for getting a store's selectors
In the previous code, 1 function is used to cover 2 use cases, switching
between them with the first parameter's type.

The original usecase is to fetch data during render. The 2nd use case
simply returns a store's selectors, so they can be called in event
handlers.

This use case does not need most of the hooks the original use is
calling. However because of how hooks work, they needed to be called
anyway in this path.

In the new function, the only remaining hook call for getting the
controls is to "useRegistry".

It also simplifies useSelect, which doesn't need any of the checks on
the type of argument anymore, and can simply return the map output.

I didn't test the impact of not having to call
these hooks, but I guess it's not 0. Especially if many components use
this.

MISSING PIECE
The function is exposed with the old signature as a public API. While in
practice it should work to check the argument and call the other
function, it's being picked up by linting. It seems safe to ignore this,
but maybe the old behavior can stay in place and deprecation is a nicer
alternative?